% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_apply.R
\name{cb_apply_filter}
\alias{cb_apply_filter}
\title{Apply a Filter}
\usage{
cb_apply_filter(
  cohort,
  filter_query,
  column_ids,
  keep_existing_filter = TRUE,
  keep_existing_columns = TRUE
)
}
\arguments{
\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}

\item{filter_query}{Phenotypic filter query.}

\item{column_ids}{Filter IDs as column in the table}

\item{keep_existing_filter}{If wants to keep existing filters (Default: TRUE)}

\item{keep_existing_columns}{If wants to keep existing columns (Default: TRUE)}
}
\value{
A confirmation string.
}
\description{
Applies filter and updates/saves the cohort from database.
}
\examples{
\dontrun{
my_cohort <- cb_load_cohort(cohort_id = "5f9af3793dd2dc6091cd17cd")
cb_apply_filter(my_cohort,
                filter_query = list("22" = list("from" = "2015-05-13", "to" = "2016-04-29"),
                                    "50" = c("Father", "Mother")) )
}

}
