% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_apply.R
\name{cb_apply_query}
\alias{cb_apply_query}
\title{Apply a query to a cohort}
\usage{
cb_apply_query(
  cohort,
  query,
  column_ids,
  keep_query = TRUE,
  keep_columns = TRUE
)
}
\arguments{
\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}

\item{query}{A phenotype query defined using the code{\link{phenotype}} function and logic operators (see example below)}

\item{column_ids}{Phenotype IDs to be added as columns in the participant table.}

\item{keep_query}{If True, combines the newly supplied query with the pre-existing query.
Otherwise, pre-existing query is overwritten. (Default: TRUE)}

\item{keep_columns}{If True, pre-existing columns are retained and newly supplied columns are added.
Otherwise, pre-exisitng columns are overwritten. (Default: TRUE)}
}
\value{
A confirmation string.
}
\description{
Updates a cohort by applying a new query.
}
