% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_cohort_create.R
\name{cb_create_cohort}
\alias{cb_create_cohort}
\title{Create Cohort}
\usage{
cb_create_cohort(cohort_name, cohort_desc, filters = "", cb_version = "v2")
}
\arguments{
\item{cohort_name}{New cohort name to be created. (Required)}

\item{cohort_desc}{New cohort description to be created. (Optional)}

\item{filters}{WIP - details will be added.}

\item{cb_version}{cohort browser version. \link{"v1" | "v2"} (Optional) Default - "v2"}
}
\value{
A \linkS4class{cohort} object.
}
\description{
Creates a new Cohort
}
\examples{
\dontrun{
my_cohort <- cb_create_cohort(cohort_name = "Cohort-R",
                              cohort_desc = "This cohort is for testing purpose, created from R.")
}
}
\seealso{
\code{\link{cb_load_cohort}} for loading a available cohort.
}
