% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_cohort_extract.R
\name{cb_extract_samples}
\alias{cb_extract_samples}
\title{Extract participants - WIP}
\usage{
cb_extract_samples(raw_data)
}
\arguments{
\item{raw_data}{A JSON string for selected participants. (Required)}
}
\value{
A dataframe.
}
\description{
Extracts selected participants.
}
