% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_explore.R
\name{cb_filter_participants}
\alias{cb_filter_participants}
\title{Filter participants}
\usage{
cb_filter_participants(cloudos, cohort, filter_id)
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}

\item{filter_id}{A filter ID. (Required)}
}
\value{
A data frame with filters applied.
}
\description{
This sums up all the filters and return number participants after applied filter.
}
