% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_explore.R
\name{cb_get_cohort_filters}
\alias{cb_get_cohort_filters}
\title{Get cohort filters}
\usage{
cb_get_cohort_filters(cohort)
}
\arguments{
\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}
}
\value{
A list of data frame.
}
\description{
Get a list of all the filters associated with a cohort.
}
