% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_explore.R
\name{cb_get_cohort_phenotypes}
\alias{cb_get_cohort_phenotypes}
\title{Get data for phenotypes associated with a cohort}
\usage{
cb_get_cohort_phenotypes(cohort)
}
\arguments{
\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}
}
\value{
A list of data frames.
}
\description{
Get a dataframe with distirbution data for each phenotype associated with a cohort.
Associated phenotypes are those found in the "Overview" section of the Cohort Browser Web UI.
}
