% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_cohort.R
\name{cb_get_filter_statistics}
\alias{cb_get_filter_statistics}
\title{Filter a cohort samples}
\usage{
cb_get_filter_statistics(cloudos, cohort, filter_id)
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}

\item{filter_id}{A filter ID. (Required)}
}
\value{
A data frame with filters applied.
}
\description{
This filters cohort samples based on particular phenotypic filter.
This will return number of samples after phenotype filter applied to a cohort.
}
