% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_cohort_extract.R
\name{cb_get_genotypic_table}
\alias{cb_get_genotypic_table}
\title{Get genotypic table}
\usage{
cb_get_genotypic_table(cloudos, cohort, size = 10, geno_filters_query)
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort}{A cohort object. (Required)
See constructor functions \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}

\item{size}{Number of entries from database. (Optional) Default - 10 (Optional)}

\item{geno_filters_query}{Genotypic filter query (Optional)}
}
\value{
A dataframe.
}
\description{
Get Genotypic table in a dataframe.
Optionally genotypic filters can be applied as well.
}
