% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_cohort_extract.R
\name{cb_get_participants_table_long}
\alias{cb_get_participants_table_long}
\title{Get longform participant data table}
\usage{
cb_get_participants_table_long(
  cohort,
  cols,
  broadcast = TRUE,
  page_number = 0,
  page_size = 100
)
}
\arguments{
\item{cohort}{A cohort object. (Required)
See constructor functions \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}

\item{cols}{Vector of phenotype IDs to fetch as columns in the dataframe. If omitted, columns saved
in the cohort are fetched.}

\item{broadcast}{Whether to broadcast single value phenotypes across rows. (Optional)
Can be TRUE, FALSE or a vector of phenotype IDs to specify which phenotypes to broadcast.
Default - TRUE}

\item{page_number}{Number of page (can be 'all' to fetch all data) . (Optional) Default - 0}

\item{page_size}{Number of entries in a page. (Optional) Default - 10}
}
\value{
A tibble.
}
\description{
Get participant data table in a longform dataframe.
}
