% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_explore.R
\name{cb_get_phenotype_metadata}
\alias{cb_get_phenotype_metadata}
\title{Phenotype metadata}
\usage{
cb_get_phenotype_metadata(pheno_id, cb_version = "v2")
}
\arguments{
\item{pheno_id}{A phenotype ID. (Required)}

\item{cb_version}{cohort browser version. (Default: "v2") \link{ "v1" | "v2" }}
}
\value{
A data frame.
}
\description{
Get the metadata of a phenotype in the cohort browser
}
