% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_class.R
\name{cb_load_cohort}
\alias{cb_load_cohort}
\title{Get cohort information}
\usage{
cb_load_cohort(cohort_id, cb_version = "v2")
}
\arguments{
\item{cohort_id}{Cohort id (Required)}

\item{cb_version}{cohort browser version (Optional) \link{ "v1" | "v2" }}
}
\value{
A \linkS4class{cohort} object.
}
\description{
Get all the details about a cohort including
applied query.
}
\seealso{
\code{\link{cb_create_cohort}} for creating a new cohort.
}
