% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_plots.R
\name{cb_plot_filters}
\alias{cb_plot_filters}
\title{Plot filters}
\usage{
cb_plot_filters(cohort)
}
\arguments{
\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}
}
\value{
A list of ggplot objects
}
\description{
Get a list of ggplot objects, each plot having one filter.
}
\examples{
\dontrun{
my_cohort <- cb_load_cohort(cohort_id = "5f9af3793dd2dc6091cd17cd")
plot_list <- cb_plot_filters(cohort = my_cohort)

plot_list[[1]]

library(ggpubr)
ggpubr::ggarrange(plotlist = plot_list)
}

}
