% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_explore.R
\name{cb_search_phenotypes}
\alias{cb_search_phenotypes}
\title{Search available phenotypes}
\usage{
cb_search_phenotypes(term, cb_version = "v2")
}
\arguments{
\item{term}{A term to search. (Required)}

\item{cb_version}{cohort browser version (Optional) [ "v1" | "v2" ]}
}
\value{
A tibble with phenotype metadata
}
\description{
Search for phenotypes in the Cohort Browser that match your term and return a tibble
containing the metadata information for each matching phenotype. Use ' term = "" ' to return all
phenotypes.
}
\examples{
\dontrun{
cancer_phenos <- cb_search_phenotypes(term = "cancer")

all_phenos <- cb_search_phenotypes(term = "")
}


}
