% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_cohort.R
\name{cb_search_phenotypic_filters}
\alias{cb_search_phenotypic_filters}
\title{List available filters}
\usage{
cb_search_phenotypic_filters(cloudos, term)
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{term}{A term to search. (Required)
Example - "cancer"}
}
\value{
A data frame with available cohort filters.
}
\description{
All the cohort filters available in CloudOS.
}
\examples{
\dontrun{
cb_search_phenotypic_filters(cloudos,
             term = "cancer")
}
}
