% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudos_class.R
\name{cloudos}
\alias{cloudos}
\title{Create a cloudos object}
\usage{
cloudos(base_url, auth, team_id)
}
\arguments{
\item{base_url}{Base URL of the CloudOS server. (Required)}

\item{auth}{An authentication method. (Required)
Example - Bearer token or API key.}

\item{team_id}{Team ID in CloudOS account. (Required)}
}
\value{
A cloudos object.
}
\description{
Creates a cloudos object which saves all login
and metadata, that can help to connect with cloudos server.
}
\examples{
\dontrun{
cloudos(base_url= "https://cloudos.lifebit.ai",
             auth = "Bearer ***token***",
             team_id = "***team_id***")
}
}
