% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudos_env.R
\name{cloudos_configure}
\alias{cloudos_configure}
\title{Configure cloudos}
\usage{
cloudos_configure(base_url, token, team_id)
}
\arguments{
\item{base_url}{Base URL for cloudos}

\item{token}{API key or token}

\item{team_id}{team/workspace ID}
}
\value{
None
}
\description{
On a system for the first time the cloudos configuration needed to be done.
This function can help do that.
}
