% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_class.R
\docType{class}
\name{cohort-class}
\alias{cohort-class}
\title{This S4 class creates a cohort object object.}
\description{
This will create a main cohort object, which will
hold many information related to a cohort.
}
\section{Slots}{

\describe{
\item{\code{cohort}}{A single cohort}

\item{\code{filter}}{A cohort filter}
}}

