% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_json.R
\name{.get_search_json}
\alias{.get_search_json}
\title{only used for v1 endpoint - creates v1 search json using the v2 style query}
\usage{
.get_search_json(my_cohort)
}
\arguments{
\item{my_cohort}{A cohort object}
}
\description{
only used for v1 endpoint - creates v1 search json using the v2 style query
}
