% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_class.R
\name{.v1_query_to_v2}
\alias{.v1_query_to_v2}
\title{Convert a v1 style query (moreFields) to v2 style (query).
v2 queries are a superset of v1 queries. A list of v1 phenotype queries are equivalent to a
set of nested v2 AND operators containing those phenotypes. This function builds the nested
AND query from the flat list of v1 phenotypes.}
\usage{
.v1_query_to_v2(cohort_more_fields)
}
\arguments{
\item{cohort_more_fields}{query information ('moreFields') from .get_cohort_info(cohort_id, cb_version="v1)}
}
\description{
Convert a v1 style query (moreFields) to v2 style (query).
v2 queries are a superset of v1 queries. A list of v1 phenotype queries are equivalent to a
set of nested v2 AND operators containing those phenotypes. This function builds the nested
AND query from the flat list of v1 phenotypes.
}
