% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_json.R
\name{.value_filter}
\alias{.value_filter}
\title{For all the helper functions here -}
\usage{
.value_filter(more_fields, fields)
}
\arguments{
\item{more_fields}{Filter information from .get_cohort_info(cohort_id)}

\item{fields}{Metainfo about filters from .get_cohort_info(cohort_id)}

\item{my_cohort}{The return list from .get_cohort_info(cohort_id)}
}
\description{
For all the helper functions here -
}
