% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_genotypic_data.R
\name{extract_genotypic_data}
\alias{extract_genotypic_data}
\title{Extract Genotypic Data}
\usage{
extract_genotypic_data(
  baseurl,
  auth,
  teamid,
  page_number = 0,
  page_size = 10,
  filters = ""
)
}
\arguments{
\item{baseurl}{Base URL of the CloudOS server. (Required)}

\item{auth}{An authentication method. (Required)
Example - Bearer token or API key.}

\item{teamid}{Team ID in CloudOS account. (Required)}

\item{page_number}{Number of page. (Optional) Default - 0}

\item{page_size}{Number of entries in a page. (Optional) Default - 10}

\item{filters}{WIP - details will be added.}
}
\value{
A dataframe.
}
\description{
Extract Genotypic Data with filter.
}
\examples{
\dontrun{
extract_genotypic_data(baseurl= "https://cloudos.lifebit.ai",
             auth = "Bearer ***token***",
             teamid = "***teamid***")
}
}
