% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_participants.R
\name{extract_participants}
\alias{extract_participants}
\title{Extract participants}
\usage{
extract_participants(base_url, auth, raw_data)
}
\arguments{
\item{base_url}{Base URL of the CloudOS server. (Required)}

\item{auth}{An authentication method. (Required)
Example - Bearer token or API key.}

\item{raw_data}{A JSON string for selected participants. (Required)}
}
\value{
A dataframe.
}
\description{
Extracts selected participants.
}
\examples{
\dontrun{
extract_participants(base_url= "https://cloudos.lifebit.ai",
             auth = "Bearer ***token***",
             raw_data = "a JSON string for selected participants")
}
}
