% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCohort.R
\name{extractCohort}
\alias{extractCohort}
\title{Cohort Dataframe Extractor}
\arguments{
\item{baseUrl}{required parameter}

\item{apikey}{required parameter}

\item{teamId}{required parameter}

\item{cohortId}{optional parameter to fetch data for a particular Id}

\item{pageNumber}{optional parameter with default value 0}

\item{pageSize}{optional parameter with default value 10}
}
\value{
output
}
\description{
Extracts the data frame with complete cohort data.
}
