% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_cohort.R
\name{filter_data}
\alias{filter_data}
\title{Filter data}
\usage{
filter_data(object, cohort_id, filter_id)
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{cloudos}}}

\item{cohort_id}{A cohort ID. (Required)}

\item{filter_id}{A filter ID. (Required)}
}
\value{
A data frame with filters applied.
}
\description{
Filters a cohort data based on particular phenotype filter.
}
