% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_cohort.R
\name{filter_metadata}
\alias{filter_metadata}
\title{Filter metadata}
\usage{
filter_metadata(object, filter_id)
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{cloudos}}}

\item{filter_id}{A filter ID. (Required)}
}
\value{
A data frame.
}
\description{
Filter metadata of a cohort filter
}
