% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_cohort.R
\name{filter_participants}
\alias{filter_participants}
\title{Filter participants}
\usage{
filter_participants(object, cohort_id, filter_id)
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{cloudos}}}

\item{cohort_id}{A cohort ID. (Required)}

\item{filter_id}{A filter ID. (Required)}
}
\value{
A data frame with filters applied.
}
\description{
This sums up all the filters and return number participants after applied filter.
}
