% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_cohort.R
\name{get_cohort_filters}
\alias{get_cohort_filters}
\title{Get cohort filters}
\usage{
get_cohort_filters(object, cohort)
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{create_cohort}} or \code{\link{load_cohort}}}
}
\value{
A list of data frame.
}
\description{
Get a list of all the filters associated with a cohort.
}
