% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_extract_cohort.R
\name{get_genotypic_table}
\alias{get_genotypic_table}
\title{Get genotypic table}
\usage{
get_genotypic_table(object, page_number = 0, page_size = 10, filters = "")
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{page_number}{Number of page. (Optional) Default - 0}

\item{page_size}{Number of entries in a page. (Optional) Default - 10}

\item{filters}{WIP - details will be added.}
}
\value{
A dataframe.
}
\description{
Get Genotypic table in a dataframe.
Optionally genotypic filters can be applied as well.
}
