% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohorts.R
\name{getCohorts}
\alias{getCohorts}
\title{Cohort List Extractor}
\arguments{
\item{baseUrl}{required parameter}

\item{apikey}{required parameter}

\item{teamId}{required parameter}

\item{cohortId}{optional parameter to fetch data for a particular Id}

\item{pageNumber}{optional parameter with default value 0}

\item{pageSize}{optional parameter with default value 10}
}
\value{
result
}
\description{
Extracts the List with complete cohort data.
}
