% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_list_cohorts.R
\name{list_cohorts}
\alias{list_cohorts}
\title{List cohorts}
\usage{
list_cohorts(object, page_number = 0, page_size = 10)
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{cloudos}}}

\item{page_number}{Number of page. (Optional) Default - 0}

\item{page_size}{Number of entries in a page. (Optional) Default - 10}
}
\value{
A data frame with available cohorts.
}
\description{
Extracts the data frame with limited cohort data columns.
}
\examples{
\dontrun{
cohorts_list(cloudos_obj)
}
}
