% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_class.R
\name{load_cohort}
\alias{load_cohort}
\title{Get cohort information}
\usage{
load_cohort(object, cohort_id)
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort_id}{Cohort id (Required)}
}
\value{
A list
}
\description{
Get all the details about a cohort including
applied filters.
}
