% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants_export.R
\name{participants_export}
\alias{participants_export}
\title{Extract participants}
\usage{
participants_export(baseurl, auth, apikey, data_raw)
}
\arguments{
\item{baseurl}{Base URL of the CloudOS server. Required}

\item{auth}{An authontication token. Example - Bearer token. Required}

\item{apikey}{A API key. Required if auth not provided.}

\item{data_raw}{A body for selected participants}
}
\value{
Ideally a dataframe.
}
\description{
Extracts selected participants.
}
