% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_plots.R
\name{plot_filters}
\alias{plot_filters}
\title{Plot filters}
\usage{
plot_filters(filter_list)
}
\arguments{
\item{filter_list}{The output of \code{\link{get_cohort_filters}} (Required)}
}
\value{
A list
}
\description{
Get a list of ggplot objects, each plot having one filter.
}
