% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_filters.R
\name{search_filters}
\alias{search_filters}
\title{List available filters}
\usage{
search_filters(base_url, auth, team_id, term)
}
\arguments{
\item{base_url}{Base URL of the CloudOS server. (Required)}

\item{auth}{An authentication method. (Required)
Example - Bearer token or API key.}

\item{team_id}{Team ID in CloudOS account. (Required)}

\item{term}{A term to search. (Required)
Example - "cancer"}
}
\value{
A data frame with available cohort filters.
}
\description{
All the cohort filters available in CloudOS.
}
\examples{
\dontrun{
search_filters(base_url= "https://cloudos.lifebit.ai",
             auth = "Bearer ***token***",
             team_id = "***team_id***",
             term = "male")
}
}
