% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_apply.R
\name{cb_apply_filter}
\alias{cb_apply_filter}
\title{Apply a Filter}
\usage{
cb_apply_filter(cloudos, cohort, filter_query)
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}

\item{filter_query}{Phenotypic filter query.}
}
\value{
A confirmation string.
}
\description{
Applies filter and updates/saves the cohort from database.
}
