% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_apply.R
\name{cb_apply_filter_dry_run}
\alias{cb_apply_filter_dry_run}
\title{Dry run for \code{\link{cb_apply_filter}}}
\usage{
cb_apply_filter_dry_run(cloudos, cohort, filter_query)
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}

\item{filter_query}{Phenotypic filter query.}
}
\value{
A data frame.
}
\description{
This doesn't update the database but mimic \code{\link{cb_apply_filter}}
}
