% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_cohort_create.R
\name{cb_create_cohort}
\alias{cb_create_cohort}
\title{Create Cohort}
\usage{
cb_create_cohort(cloudos, cohort_name, cohort_desc, filters = "")
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort_name}{New cohort name to be created. (Required)}

\item{cohort_desc}{New cohort description to be created. (Optional)}

\item{filters}{WIP - details will be added.}
}
\value{
A \linkS4class{cohort} object.
}
\description{
Creates a new Cohort
}
\examples{
\dontrun{
cb_create_cohort(cloudos_obj,
             cohort_name = "my cohort",
             cohort_desc = "my cohort description")
}
}
\seealso{
\code{\link{cb_load_cohort}} for loading a available cohort.
}
