% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_extract_cohort.R
\name{cb_extract_samples}
\alias{cb_extract_samples}
\title{Extract participants}
\usage{
cb_extract_samples(object, raw_data)
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{raw_data}{A JSON string for selected participants. (Required)}
}
\value{
A dataframe.
}
\description{
Extracts selected participants.
}
