% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_explore.R
\name{cb_filter_metadata}
\alias{cb_filter_metadata}
\title{Filter metadata}
\usage{
cb_filter_metadata(cloudos, filter_id)
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{filter_id}{A filter ID. (Required)}
}
\value{
A data frame.
}
\description{
Filter metadata of a cohort filter
}
