% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_cohort_extract.R
\name{cb_get_genotypic_table}
\alias{cb_get_genotypic_table}
\title{Get genotypic table}
\usage{
cb_get_genotypic_table(cloudos, page_number = 0, page_size = 10, filters = "")
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{page_number}{Number of page. (Optional) Default - 0}

\item{page_size}{Number of entries in a page. (Optional) Default - 10}

\item{filters}{WIP - details will be added.}
}
\value{
A dataframe.
}
\description{
Get Genotypic table in a dataframe.
Optionally genotypic filters can be applied as well.
}
