% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_cohort_extract.R
\name{cb_get_participants_table}
\alias{cb_get_participants_table}
\title{Get participant data table}
\usage{
cb_get_participants_table(cohort, cols, page_number = "all", page_size = 5000)
}
\arguments{
\item{cohort}{A cohort object. (Required)
See constructor functions \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}

\item{cols}{Vector of phenotype IDs to fetch as columns in the dataframe. If omitted, columns saved
in the cohort are fetched.}

\item{page_number}{Number of page as integer or 'all' to fetch all data. (Optional) Default - 'all'}

\item{page_size}{Number of entries in a page. (Optional) Default - 5000}
}
\value{
A dataframe.
}
\description{
Get participant data table in a dataframe.
}
