% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_cohorts_list.R
\name{cb_list_cohorts}
\alias{cb_list_cohorts}
\title{List cohorts}
\usage{
cb_list_cohorts(size = 10)
}
\arguments{
\item{size}{Number of cohort entries from database. (Optional) Default - 10}
}
\value{
A data frame with available cohorts.
}
\description{
Extracts the data frame with limited cohort data columns.
}
\examples{
\dontrun{
cohorts_list()
}
}
