% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_class.R
\name{cb_load_cohort}
\alias{cb_load_cohort}
\title{Get cohort information}
\usage{
cb_load_cohort(cloudos, cohort_id)
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort_id}{Cohort id (Required)}
}
\value{
A \linkS4class{cohort} object.
}
\description{
Get all the details about a cohort including
applied filters.
}
\seealso{
\code{\link{cb_create_cohort}} for creating a new cohort.
}
