% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_explore.R
\name{cb_participant_count}
\alias{cb_participant_count}
\title{Participant Count}
\usage{
cb_participant_count(cohort, query = list(), keep_query = TRUE)
}
\arguments{
\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}

\item{query}{A phenotype query defined using the code{\link{phenotype}} function and logic operators (see example below)}

\item{keep_query}{Apply newly specified query on top of exisiting query (Default: TRUE)}
}
\value{
A list with count of participants in the cohort and the total no. of participants in the dataset.
}
\description{
Returns the number of participants in a cohort if the supplied query were to be applied.
}
