% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_plots.R
\name{cb_plot_filters}
\alias{cb_plot_filters}
\title{Plot filters}
\usage{
cb_plot_filters(cloudos, cohort)
}
\arguments{
\item{cloudos}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}}
}
\value{
A list of ggplot objects
}
\description{
Get a list of ggplot objects, each plot having one filter.
}
