% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_explore.R
\name{cb_search_phenotypic_filters}
\alias{cb_search_phenotypic_filters}
\title{List available filters}
\usage{
cb_search_phenotypic_filters(term)
}
\arguments{
\item{term}{A term to search. (Required)
Example - "cancer"}
}
\value{
A data frame with available cohort filters.
}
\description{
All the cohort filters available in CloudOS.
}
\examples{
\dontrun{
cb_search_phenotypic_filters(term = "cancer")
}
}
