% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudos_class.R
\docType{class}
\name{cloudos-class}
\alias{cloudos-class}
\title{cloudos class}
\description{
This will create a main object, which will
hold many information related to CloudOS server
including base URL, authentication and team ID.
Use of this object will simplify calling every
other functions which requires to communicate with server.
}
\section{Slots}{

\describe{
\item{\code{base_url}}{Base URL of the CloudOS server.}

\item{\code{auth}}{An authentication method.}

\item{\code{auth_method}}{Which authentication method being used.}

\item{\code{team_id}}{Team ID in CloudOS account.}
}}

