% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_class.R
\docType{class}
\name{cohort-class}
\alias{cohort-class}
\title{cohort class}
\description{
This will create a cohort object, which will
hold many information related to a cohort
including cohort ID, name, description, filters, tables.
Use of this object will simplify calling every
other functions which requires a cohort.
}
\section{Slots}{

\describe{
\item{\code{id}}{cohort ID.}

\item{\code{name}}{cohort name.}

\item{\code{desc}}{cohort description.}

\item{\code{fields}}{Filters.}

\item{\code{more_fields}}{Filter related information.}
}}

