% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_class.R
\docType{class}
\name{cohort-class}
\alias{cohort-class}
\title{cohort class}
\description{
This class creates a cohort object, which holds the information related to a
cohort: cohort ID, name, description, query, table columns. This class is used
in functions which carry out operations related to specific cohorts.
A cohort class object can be created using constructor functions
\code{\link{cb_create_cohort}} or \code{\link{cb_load_cohort}}.
}
\section{Slots}{

\describe{
\item{\code{id}}{cohort ID.}

\item{\code{name}}{cohort name.}

\item{\code{desc}}{cohort description.}

\item{\code{phenoptype_filters}}{phenotypes displayed in the cohort overview.}

\item{\code{query}}{applied query.}

\item{\code{query_phenotype_ids}}{IDs of phenotypes used in the query.}

\item{\code{columns}}{All the columns.}

\item{\code{num_participants}}{number of participants in the cohort.}

\item{\code{cb_version}}{chort browser version.}
}}

