% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort.R
\name{create_cohort}
\alias{create_cohort}
\title{Create Cohort}
\usage{
create_cohort(base_url, auth, team_id, cohort_name, cohort_desc, filters = "")
}
\arguments{
\item{base_url}{Base URL of the CloudOS server. (Required)}

\item{auth}{An authentication method. (Required)
Example - Bearer token or API key.}

\item{team_id}{Team ID in CloudOS account. (Required)}

\item{cohort_name}{New cohort name to be created. (Required)}

\item{cohort_desc}{New cohort description to be created. (Required)}

\item{filters}{WIP - details will be added.}
}
\value{
A dataframe.
}
\description{
Creates a new Cohort
}
\examples{
\dontrun{
create_cohort(base_url= "https://cloudos.lifebit.ai",
             auth = "Bearer ***token***",
             team_id = "***team_id***",
             cohort_name = "my cohort",
             cohort_desc = "my cohort description")
}
}
