% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_create_cohort.R
\name{create_cohort}
\alias{create_cohort}
\title{Create Cohort}
\usage{
create_cohort(object, cohort_name, cohort_desc, filters = "")
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort_name}{New cohort name to be created. (Required)}

\item{cohort_desc}{New cohort description to be created. (Optional)}

\item{filters}{WIP - details will be added.}
}
\value{
A dataframe.
}
\description{
Creates a new Cohort
}
\examples{
\dontrun{
create_cohort(cloudos_obj,
             cohort_name = "my cohort",
             cohort_desc = "my cohort description")
}
}
