% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_extract_cohort.R
\name{extract_genotypic_data}
\alias{extract_genotypic_data}
\title{Extract Genotypic Data}
\usage{
extract_genotypic_data(object, page_number = 0, page_size = 10, filters = "")
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{cloudos}}}

\item{page_number}{Number of page. (Optional) Default - 0}

\item{page_size}{Number of entries in a page. (Optional) Default - 10}

\item{filters}{WIP - details will be added.}
}
\value{
A dataframe.
}
\description{
Extract Genotypic Data with filter.
}
\examples{
\dontrun{
extract_genotypic_data(cloudos_obj)
}
}
