% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_participants.R
\name{extract_participants}
\alias{extract_participants}
\title{Extract participants}
\usage{
extract_participants(baseurl, auth, apikey, raw_data)
}
\arguments{
\item{baseurl}{Base URL of the CloudOS server. Required}

\item{auth}{An authontication token. Example - Bearer token. Required}

\item{apikey}{A API key. Required if auth not provided.}

\item{raw_data}{A JSON string for selected participants.}
}
\value{
Ideally a dataframe.
}
\description{
Extracts selected participants.
}
