% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_cohort.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter a cohort samples}
\usage{
filter_samples(object, cohort, filter_id)
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cohort}}}

\item{filter_id}{A filter ID. (Required)}
}
\value{
A data frame with filters applied.
}
\description{
This filters cohort samples based on particular phenotypic filter.
This will return number of samples after phenotype filter applied to a cohort.
}
