% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_cohort.R
\name{genotypic_save}
\alias{genotypic_save}
\title{Genotypic save}
\usage{
genotypic_save(object, cohort, filter_id)
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{cohort}{A cohort object. (Required)
See constructor function \code{\link{cohort}}}

\item{filter_id}{A filter ID. (Required)}
}
\value{
A data frame.
}
\description{
applies filter (genotypic-save). Returns df with cohort and filtered participants
}
