% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_filter_cohort.R
\name{search_filters}
\alias{search_filters}
\title{List available filters}
\usage{
search_filters(object, term)
}
\arguments{
\item{object}{A cloudos object. (Required)
See constructor function \code{\link{connect_cloudos}}}

\item{term}{A term to search. (Required)
Example - "cancer"}
}
\value{
A data frame with available cohort filters.
}
\description{
All the cohort filters available in CloudOS.
}
\examples{
\dontrun{
search_filters(cloudos_object,
             term = "male")
}
}
